fix integers, move logic to planning service, add assigning referees to matches

+++ backend/src/main/java/com/roc/app/competition/referee/dto/CompetitionRefereeResponseDto.java
@@ -2,7 +2,7 @@ package com.roc.app.competition.referee.dto;
 
 import com.roc.app.competition.referee.CompetitionReferee;
 
-public record CompetitionRefereeResponseDto(Integer competitionId, Long refereeId) {
+public record CompetitionRefereeResponseDto(Integer competitionId, Integer refereeId) {
         public static CompetitionRefereeResponseDto fromModel(CompetitionReferee model) {
             return new CompetitionRefereeResponseDto(
                     model.getCompetitionId(),
